feat: change file struct
