+++ pkg/installer/oneagent/verify.go
@@ -1,4 +1,4 @@
-package installer
+package oneagent
 
 import (
 	"bytes"
